requirements.txt
atomicwrites==1.3.0
attrs==18.2.0
colorama==0.4.1
more-itertools==5.0.0
pluggy==0.8.1
py==1.7.0
pytest==4.2.0
selenium==3.141.0
six==1.12.0
urllib3==1.24.1
allure-pytest==2.5.5
allure-python-commons==2.5.5
atomicwrites==1.3.0
attrs==18.2.0
colorama==0.4.1
future==0.17.1
more-itertools==5.0.0
pluggy==0.8.1
py==1.7.0
pytest==4.2.0
selene==0.0.8
selenium==2.53.1
six==1.12.0
urllib3==1.24.1
allure-pytest==2.5.5
allure-python-commons==2.5.5
atomicwrites==1.3.0
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
contextlib2==0.5.5
docopt==0.6.2
future==0.17.1
idna==2.8
jmespath==0.9.3
more-itertools==5.0.0
paho-mqtt==1.3.1
pbr==5.1.2
pluggy==0.8.1
py==1.7.0
PyJWT==1.7.1
pykwalify==1.7.0
pytest==4.2.0
pytest-html==1.20.0
pytest-metadata==1.8.0
python-box==3.2.4
python-dateutil==2.8.0
PyYAML==3.13
requests==2.21.0
selenium==3.141.0
six==1.12.0
stevedore==1.30.0
tavern==0.22.1
urllib3==1.24.1
